+++ src/rex.deploy/NEWS.rst
@@ -10,6 +10,7 @@
 
 * Added support for psycopg2 >= 2.5.
 * Support for serializing facts in YAML.
+* Documentation for model entities.
 
 
 2.4.1 (2016-01-29)